+++ CHANGELOG.md
@@ -2,8 +2,12 @@
 
 ## 3.0.0-beta.3 - Unreleased
 ### Added
-- Added `commandsBefore` and `commandsAfter` config settings to Git Deployer.
-- Added `EVENT_BEFORE_COMMIT` and `EVENT_AFTER_COMMIT` events to Git Deployer.
+- Added the `refreshCacheAutomaticallyForGlobals` config setting.
+- Added the `commandsBefore` and `commandsAfter` config settings to Git Deployer.
+- Added the `EVENT_BEFORE_COMMIT` and `EVENT_AFTER_COMMIT` events to Git Deployer.
+
+### Changed
+- Removed the `clearCacheAutomaticallyForGlobals` and `warmCacheAutomaticallyForGlobals` config settings.
 
 ## 3.0.0-beta.2.1 - 2019-10-29
 ### Fixed